Refresh libctru to the latest version

+++ ctru-rs/src/services/ps.rs
@@ -46,7 +46,7 @@ pub fn device_id() -> crate::Result<u32> {
 
 pub fn generate_random_bytes(out: &mut [u8]) -> crate::Result<()> {
     ResultCode(unsafe {
-        ctru_sys::PS_GenerateRandomBytes(out as *mut _ as *mut _, out.len() as u32)
+        ctru_sys::PS_GenerateRandomBytes(out as *mut _ as *mut _, out.len())
     })?;
     Ok(())
 }
